backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.40.0",
    "google-generativeai>=0.8.0",
    "httpx>=0.27.0",
]
